+++ lib/processor/log.go
@@ -1,11 +1,15 @@
 package processor
 
 import (
+	"errors"
 	"fmt"
+	"sort"
 	"time"
 
 	"github.com/Jeffail/benthos/v3/internal/bloblang"
 	"github.com/Jeffail/benthos/v3/internal/bloblang/field"
+	"github.com/Jeffail/benthos/v3/internal/bloblang/mapping"
+	"github.com/Jeffail/benthos/v3/internal/bloblang/query"
 	"github.com/Jeffail/benthos/v3/internal/docs"
 	"github.com/Jeffail/benthos/v3/lib/log"
 	"github.com/Jeffail/benthos/v3/lib/metrics"
@@ -48,24 +52,46 @@ can be any of the following values: TRACE, DEBUG, INFO, WARN, ERROR.
 
 ### Structured Fields
 
-It's also possible to output a map of structured fields, this only works when
-the service log is set to output as JSON. The field values are function
-interpolated, meaning it's possible to output structured fields containing
-message contents and metadata, e.g.:
+It's also possible add custom fields to logs when the format is set to a structured form such as ` + "`json` or `logfmt`" + `. The config field ` + "`fields`" + ` allows you to provide a map of key/value string pairs, where the values support [interpolation functions](/docs/configuration/interpolation#bloblang-queries) allowing you to extract message contents and metadata like this:
 
 ` + "```yaml" + `
 pipeline:
   processors:
     - log:
         level: DEBUG
-        message: "foo"
+        message: hello world
         fields:
-          id: '${! json("id") }'
-          kafka_topic: '${! meta("kafka_topic") }'
-` + "```" + ``,
+          reason: cus I wana
+          id: ${! json("id") }
+          age: ${! json("user.age") }
+          kafka_topic: ${! meta("kafka_topic") }
+` + "```" + `
+
+However, these values will always be output as string types. In cases where you want to add other types such as integers or booleans you can use the field ` + "`fields_mapping`" + ` to define a [Bloblang mapping](/docs/guides/bloblang/about) that outputs a map of key/values like this:
+
+` + "```yaml" + `
+pipeline:
+  processors:
+    - log:
+        level: DEBUG
+        message: hello world
+        fields_mapping: |
+          root.reason = "cus I wana"
+          root.id = this.id
+          root.age = this.user.age
+          root.kafka_topic = meta("kafka_topic")
+` + "```" + `
+`,
 		FieldSpecs: docs.FieldSpecs{
 			docs.FieldCommon("level", "The log level to use.").HasOptions("FATAL", "ERROR", "WARN", "INFO", "DEBUG", "TRACE", "ALL"),
 			docs.FieldCommon("fields", "A map of fields to print along with the log message.").SupportsInterpolation(true),
+			docs.FieldCommon(
+				"fields_mapping", "An optional [Bloblang mapping](/docs/guides/bloblang/about) that can be used to specify extra fields to add to the log. If log fields are also added with the `fields` field then those values will override matching keys from this mapping.",
+				`root.reason = "cus I wana"
+root.id = this.id
+root.age = this.user.age.number()
+root.kafka_topic = meta("kafka_topic")`,
+			).AtVersion("3.40.0"),
 			docs.FieldCommon("message", "The message to print.").SupportsInterpolation(true),
 		},
 	}
@@ -75,29 +101,39 @@ pipeline:
 
 // LogConfig contains configuration fields for the Log processor.
 type LogConfig struct {
-	Level   string            `json:"level" yaml:"level"`
-	Fields  map[string]string `json:"fields" yaml:"fields"`
-	Message string            `json:"message" yaml:"message"`
+	Level         string            `json:"level" yaml:"level"`
+	Fields        map[string]string `json:"fields" yaml:"fields"`
+	FieldsMapping string            `json:"fields_mapping" yaml:"fields_mapping"`
+	Message       string            `json:"message" yaml:"message"`
 }
 
 // NewLogConfig returns a LogConfig with default values.
 func NewLogConfig() LogConfig {
 	return LogConfig{
-		Level:   "INFO",
-		Fields:  map[string]string{},
-		Message: "",
+		Level:         "INFO",
+		Fields:        map[string]string{},
+		FieldsMapping: "",
+		Message:       "",
 	}
 }
 
 //------------------------------------------------------------------------------
 
+type logWith interface {
+	log.Modular
+	With(args ...interface{}) log.Modular
+}
+
 // Log is a processor that prints a log event each time it processes a message.
 type Log struct {
-	log     log.Modular
+	logger  log.Modular
 	level   string
 	message field.Expression
 	fields  map[string]field.Expression
 	printFn func(logger log.Modular, msg string)
+
+	loggerWith    logWith
+	fieldsMapping *mapping.Executor
 }
 
 // NewLog returns a Log processor.
@@ -109,7 +145,7 @@ func NewLog(
 		return nil, fmt.Errorf("failed to parse message expression: %v", err)
 	}
 	l := &Log{
-		log:     logger,
+		logger:  logger,
 		level:   conf.Log.Level,
 		fields:  map[string]field.Expression{},
 		message: message,
@@ -119,6 +155,15 @@ func NewLog(
 			if l.fields[k], err = bloblang.NewField(v); err != nil {
 				return nil, fmt.Errorf("failed to parse field '%v' expression: %v", k, err)
 			}
+		}
+	}
+	if len(conf.Log.FieldsMapping) > 0 {
+		var ok bool
+		if l.loggerWith, ok = logger.(logWith); !ok {
+			return nil, errors.New("the provided logger does not support structured fields required for `fields_mapping`")
+		}
+		if l.fieldsMapping, err = bloblang.NewMapping("", conf.Log.FieldsMapping); err != nil {
+			return nil, fmt.Errorf("failed to parse fields mapping: %w", err)
 		}
 	}
 	if l.printFn, err = l.levelToLogFn(l.level); err != nil {
@@ -159,7 +204,54 @@ func (l *Log) levelToLogFn(level string) (func(logger log.Modular, msg string), 
 
 // ProcessMessage logs an event and returns the message unchanged.
 func (l *Log) ProcessMessage(msg types.Message) ([]types.Message, types.Response) {
-	targetLog := l.log
+	targetLog := l.logger
+	if l.fieldsMapping != nil {
+		v, err := l.fieldsMapping.Exec(query.FunctionContext{
+			Maps:     map[string]query.Function{},
+			Vars:     map[string]interface{}{},
+			Index:    0,
+			MsgBatch: msg,
+		}.WithValueFunc(func() *interface{} {
+			jObj, err := msg.Get(0).JSON()
+			if err != nil {
+				return nil
+			}
+			return &jObj
+		}))
+		if err != nil {
+			l.logger.Errorf("Failed to execute fields mapping: %v", err)
+			resMsg := msg.Copy()
+			resMsg.Iter(func(i int, p types.Part) error {
+				FlagErr(p, err)
+				return nil
+			})
+			return []types.Message{resMsg}, nil
+		}
+		vObj, ok := v.(map[string]interface{})
+		if !ok {
+			l.logger.Errorf("Fields mapping yielded a non-object result: %T", v)
+			rErr := fmt.Errorf("fields mapping yielded a non-object result: %T", v)
+			resMsg := msg.Copy()
+			resMsg.Iter(func(i int, p types.Part) error {
+				FlagErr(p, rErr)
+				return nil
+			})
+			return []types.Message{resMsg}, nil
+		}
+
+		keys := make([]string, 0, len(vObj))
+		for k := range vObj {
+			keys = append(keys, k)
+		}
+		sort.Strings(keys)
+
+		args := make([]interface{}, 0, len(vObj)*2)
+		for _, k := range keys {
+			args = append(args, k)
+			args = append(args, vObj[k])
+		}
+		targetLog = l.loggerWith.With(args...)
+	}
 	if len(l.fields) > 0 {
 		interpFields := make(map[string]string, len(l.fields))
 		for k, vi := range l.fields {
@@ -167,9 +259,8 @@ func (l *Log) ProcessMessage(msg types.Message) ([]types.Message, types.Response
 		}
 		targetLog = log.WithFields(targetLog, interpFields)
 	}
-	msgs := [1]types.Message{msg}
 	l.printFn(targetLog, l.message.String(0, msg))
-	return msgs[:], nil
+	return []types.Message{msg}, nil
 }
 
 // CloseAsync shuts down the processor and stops processing requests.